Fix horizon validation for non-integer angle divisions (#665)
Fixes #647 - Change azimuthFrom/azimuthTo to float type to handle horizon counts that don't divide evenly into 360°

+++ src/akkudoktoreos/prediction/pvforecastakkudoktor.py
@@ -92,8 +92,8 @@ from akkudoktoreos.utils.datetimeutil import compare_datetimes, to_datetime
 
 class AkkudoktorForecastHorizon(PydanticBaseModel):
     altitude: int
-    azimuthFrom: int
-    azimuthTo: int
+    azimuthFrom: float
+    azimuthTo: float
 
 
 class AkkudoktorForecastMeta(PydanticBaseModel):
